services/diagnostic-services/look-service/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.2
pydantic-settings==2.1.0
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-health-checking==1.54.0
grpcio-reflection==1.54.0
protobuf==4.25.1
PyYAML==6.0.1

# Database & Storage
motor==3.3.1
pymongo==4.3.3
redis==4.5.4
pymilvus==2.3.3

# ML & Computer Vision
torch==2.0.0
torchvision==0.15.1
onnx==1.15.0
onnxruntime-gpu==1.16.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-image==0.19.3
mediapipe==0.10.5
dlib==19.24.2
albumentations==1.3.1
numpy==1.26.2

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
python-logging-loki==0.3.1
python-dotenv==1.0.0
python-json-logger==2.0.7

# Utilities
python-multipart==0.0.7
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
structlog==23.2.0
retry==0.9.2
tqdm==4.66.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.0
isort==5.13.2
mypy==1.7.1
flake8==6.1.0
pylint==3.0.3

# Additional dependencies
boto3==1.26.115

# Service discovery and health checking
python-consul==1.1.0
etcd3==0.12.0
psutil==5.9.5

# Pybreaker for circuit breaking
pybreaker==0.7.0

# Database
SQLAlchemy==2.0.25 